Break chain_client and zetacore_observer into smaller files
revert logic changes
