+++ zetaclient/chain_client.go
@@ -2,12 +2,9 @@ package zetaclient
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/binary"
-	"errors"
 	"fmt"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/syndtr/goleveldb/leveldb/util"
 	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 	"math/big"
@@ -17,19 +14,16 @@ import (
 	"sync"
 	"time"
 
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/zeta-chain/zetacore/common"
-	"github.com/zeta-chain/zetacore/zetaclient/config"
-	"github.com/zeta-chain/zetacore/zetaclient/types"
-
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
-	zetatypes "github.com/zeta-chain/zetacore/x/zetacore/types"
+	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
+	"github.com/syndtr/goleveldb/leveldb"
+	"github.com/zeta-chain/zetacore/common"
+	"github.com/zeta-chain/zetacore/zetaclient/config"
 )
 
 const (
@@ -94,8 +88,8 @@ type ChainObserver struct {
 	endpoint         string
 	ticker           *time.Ticker
 	connectorAbi     *abi.ABI // token contract ABI on non-ethereum chain; zetalocker on ethereum
-	Client           *ethclient.Client
-	bridge           *ZetaCoreBridge
+	EvmClient        *ethclient.Client
+	zetaClient       *ZetaCoreBridge
 	Tss              TSSSigner
 	LastBlock        uint64
 	confCount        uint64 // must wait this many blocks to be considered "confirmed"
@@ -105,8 +99,8 @@ type ChainObserver struct {
 	db               *leveldb.DB
 	sampleLogger     *zerolog.Logger
 	metrics          *metrics.Metrics
-	nonceTxHashesMap map[int][]string
-	nonceTx          map[int]*ethtypes.Receipt
+	outTXPending     map[int][]string
+	outTXConfirmed   map[int]*ethtypes.Receipt
 	MinNonce         int
 	MaxNonce         int
 	OutTxChan        chan OutTx // send to this channel if you want something back!
@@ -125,12 +119,12 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 	ob.mu = &sync.Mutex{}
 	sampled := log.Sample(&zerolog.BasicSampler{N: 10})
 	ob.sampleLogger = &sampled
-	ob.bridge = bridge
+	ob.zetaClient = bridge
 	ob.txWatchList = make(map[ethcommon.Hash]string)
 	ob.Tss = tss
 	ob.metrics = metrics
-	ob.nonceTxHashesMap = make(map[int][]string)
-	ob.nonceTx = make(map[int]*ethtypes.Receipt)
+	ob.outTXPending = make(map[int][]string)
+	ob.outTXConfirmed = make(map[int]*ethtypes.Receipt)
 	ob.OutTxChan = make(chan OutTx, 100)
 	ob.mpiAddress = config.Chains[chain.String()].ConnectorContractAddress
 	ob.endpoint = config.Chains[chain.String()].Endpoint
@@ -150,7 +144,7 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 		log.Err(err).Msg("eth Client Dial")
 		return nil, err
 	}
-	ob.Client = client
+	ob.EvmClient = client
 
 	// create metric counters
 	err = ob.RegisterPromCounter("rpc_getLogs_count", "Number of getLogs")
@@ -180,19 +174,19 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 	// initialize zeta price queriers
 	uniswapv3querier := &UniswapV3ZetaPriceQuerier{
 		UniswapV3Abi:        &uniswapV3ABI,
-		Client:              ob.Client,
+		Client:              ob.EvmClient,
 		PoolContractAddress: ethcommon.HexToAddress(config.Chains[chain.String()].PoolContractAddress),
 		Chain:               ob.chain,
 	}
 	uniswapv2querier := &UniswapV2ZetaPriceQuerier{
 		UniswapV2Abi:        &uniswapV2ABI,
-		Client:              ob.Client,
+		Client:              ob.EvmClient,
 		PoolContractAddress: ethcommon.HexToAddress(config.Chains[chain.String()].PoolContractAddress),
 		Chain:               ob.chain,
 	}
 	dummyQuerier := &DummyZetaPriceQuerier{
 		Chain:  ob.chain,
-		Client: ob.Client,
+		Client: ob.EvmClient,
 	}
 
 	// Initialize chain specific setup
@@ -239,7 +233,7 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 		envvar := ob.chain.String() + "_SCAN_CURRENT"
 		if os.Getenv(envvar) != "" {
 			log.Info().Msgf("envvar %s is set; scan from current block", envvar)
-			header, err := ob.Client.HeaderByNumber(context.Background(), nil)
+			header, err := ob.EvmClient.HeaderByNumber(context.Background(), nil)
 			if err != nil {
 				return nil, err
 			}
@@ -251,7 +245,7 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 				ob.LastBlock = ob.getLastHeight()
 				// if ZetaCore does not have last heard block height, then use current
 				if ob.LastBlock == 0 {
-					header, err := ob.Client.HeaderByNumber(context.Background(), nil)
+					header, err := ob.EvmClient.HeaderByNumber(context.Background(), nil)
 					if err != nil {
 						return nil, err
 					}
@@ -278,7 +272,7 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 					continue
 				}
 				txHashes := strings.Split(string(iter.Value()), ",")
-				ob.nonceTxHashesMap[int(nonce)] = txHashes
+				ob.outTXPending[int(nonce)] = txHashes
 				log.Info().Msgf("reading nonce %d with %d tx hashes", nonce, len(txHashes))
 			}
 			iter.Release()
@@ -302,7 +296,7 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 					log.Error().Err(err).Msgf("error unmarshalling receipt: %s", key)
 					continue
 				}
-				ob.nonceTx[int(nonce)] = &receipt
+				ob.outTXConfirmed[int(nonce)] = &receipt
 				log.Info().Msgf("chain %s reading nonce %d with receipt of tx %s", ob.chain, nonce, receipt.TxHash.Hex())
 			}
 			iter.Release()
@@ -320,351 +314,11 @@ func NewChainObserver(chain common.Chain, bridge *ZetaCoreBridge, tss TSSSigner,
 	return &ob, nil
 }
 
-func (ob *ChainObserver) GetPromCounter(name string) (prometheus.Counter, error) {
-	if cnt, found := metrics.Counters[ob.chain.String()+"_"+name]; found {
-		return cnt, nil
-	} else {
-		return nil, errors.New("counter not found")
-	}
-}
-
-func (ob *ChainObserver) RegisterPromCounter(name string, help string) error {
-	cntName := ob.chain.String() + "_" + name
-	return ob.metrics.RegisterCounter(cntName, help)
-}
-
 func (ob *ChainObserver) Start() {
-	go ob.WatchRouter()
-	go ob.WatchGasPrice()
-	go ob.WatchExchangeRate()
+	go ob.ExternalChainWatcher() // Observes external Chains for incoming trasnactions
+	go ob.WatchGasPrice()        // Observes external Chains for Gas prices and posts to core
+	go ob.WatchExchangeRate()    // Observers ZetaPriceQuerier for Zeta prices and posts to core
 	go ob.observeOutTx()
-}
-
-func (ob *ChainObserver) PostNonceIfNotRecorded() error {
-	bridge := ob.bridge
-	client := ob.Client
-	tss := ob.Tss
-	chain := ob.chain
-
-	_, err := bridge.GetNonceByChain(chain)
-	if err != nil { // if Nonce of Chain is not found in ZetaCore; report it
-		nonce, err := client.NonceAt(context.TODO(), tss.Address(), nil)
-		if err != nil {
-			log.Err(err).Msg("NonceAt")
-			return err
-		}
-		log.Debug().Msgf("signer %s Posting Nonce of chain %s of nonce %d", bridge.GetKeys().signerName, chain, nonce)
-		_, err = bridge.PostNonce(chain, nonce)
-		if err != nil {
-			log.Err(err).Msg("PostNonce")
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (ob *ChainObserver) WatchRouter() {
-	// At each tick, query the mpiAddress
-	for {
-		select {
-		case <-ob.ticker.C:
-			err := ob.observeChain()
-			if err != nil {
-				log.Err(err).Msg("observeChain error on " + ob.chain.String())
-				continue
-			}
-		case <-ob.stop:
-			log.Info().Msg("WatchRouter stopped")
-			return
-		}
-	}
-}
-
-func (ob *ChainObserver) WatchGasPrice() {
-	gasTicker := time.NewTicker(60 * time.Second)
-	for {
-		select {
-		case <-gasTicker.C:
-			err := ob.PostGasPrice()
-			if err != nil {
-				log.Err(err).Msg("PostGasPrice error on " + ob.chain.String())
-				continue
-			}
-		case <-ob.stop:
-			log.Info().Msg("WatchGasPrice stopped")
-			return
-		}
-	}
-}
-
-func (ob *ChainObserver) WatchExchangeRate() {
-	ticker := time.NewTicker(60 * time.Second)
-	for {
-		select {
-		case <-ticker.C:
-			price, bn, err := ob.ZetaPriceQuerier.GetZetaPrice()
-			if err != nil {
-				log.Err(err).Msg("GetZetaExchangeRate error on " + ob.chain.String())
-				continue
-			}
-			priceInHex := fmt.Sprintf("0x%x", price)
-
-			_, err = ob.bridge.PostZetaConversionRate(ob.chain, priceInHex, bn)
-			if err != nil {
-				log.Err(err).Msg("PostZetaConversionRate error on " + ob.chain.String())
-			}
-		case <-ob.stop:
-			log.Info().Msg("WatchExchangeRate stopped")
-			return
-		}
-	}
-}
-
-func (ob *ChainObserver) PostGasPrice() error {
-	// GAS PRICE
-	gasPrice, err := ob.Client.SuggestGasPrice(context.TODO())
-	if err != nil {
-		log.Err(err).Msg("PostGasPrice:")
-		return err
-	}
-	blockNum, err := ob.Client.BlockNumber(context.TODO())
-	if err != nil {
-		log.Err(err).Msg("PostGasPrice:")
-		return err
-	}
-
-	// SUPPLY
-	var supply string // lockedAmount on ETH, totalSupply on other chains
-	supply = "100"
-	//if chainOb.chain == common.ETHChain {
-	//	input, err := chainOb.connectorAbi.Pack("getLockedAmount")
-	//	if err != nil {
-	//		return fmt.Errorf("fail to getLockedAmount")
-	//	}
-	//	bn, err := chainOb.Client.BlockNumber(context.TODO())
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s BlockNumber error", chainOb.chain)
-	//		return err
-	//	}
-	//	fromAddr := ethcommon.HexToAddress(config.TSS_TEST_ADDRESS)
-	//	toAddr := ethcommon.HexToAddress(config.ETH_MPI_ADDRESS)
-	//	res, err := chainOb.Client.CallContract(context.TODO(), ethereum.CallMsg{
-	//		From: fromAddr,
-	//		To:   &toAddr,
-	//		Data: input,
-	//	}, big.NewInt(0).SetUint64(bn))
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s CallContract error", chainOb.chain)
-	//		return err
-	//	}
-	//	output, err := chainOb.connectorAbi.Unpack("getLockedAmount", res)
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s Unpack error", chainOb.chain)
-	//		return err
-	//	}
-	//	lockedAmount := *connectorAbi.ConvertType(output[0], new(*big.Int)).(**big.Int)
-	//	//fmt.Printf("ETH: block %d: lockedAmount %d\n", bn, lockedAmount)
-	//	supply = lockedAmount.String()
-	//
-	//} else if chainOb.chain == common.BSCChain {
-	//	input, err := chainOb.connectorAbi.Pack("totalSupply")
-	//	if err != nil {
-	//		return fmt.Errorf("fail to totalSupply")
-	//	}
-	//	bn, err := chainOb.Client.BlockNumber(context.TODO())
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s BlockNumber error", chainOb.chain)
-	//		return err
-	//	}
-	//	fromAddr := ethcommon.HexToAddress(config.TSS_TEST_ADDRESS)
-	//	toAddr := ethcommon.HexToAddress(config.BSC_MPI_ADDRESS)
-	//	res, err := chainOb.Client.CallContract(context.TODO(), ethereum.CallMsg{
-	//		From: fromAddr,
-	//		To:   &toAddr,
-	//		Data: input,
-	//	}, big.NewInt(0).SetUint64(bn))
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s CallContract error", chainOb.chain)
-	//		return err
-	//	}
-	//	output, err := chainOb.connectorAbi.Unpack("totalSupply", res)
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s Unpack error", chainOb.chain)
-	//		return err
-	//	}
-	//	totalSupply := *connectorAbi.ConvertType(output[0], new(*big.Int)).(**big.Int)
-	//	//fmt.Printf("BSC: block %d: totalSupply %d\n", bn, totalSupply)
-	//	supply = totalSupply.String()
-	//} else if chainOb.chain == common.POLYGONChain {
-	//	input, err := chainOb.connectorAbi.Pack("totalSupply")
-	//	if err != nil {
-	//		return fmt.Errorf("fail to totalSupply")
-	//	}
-	//	bn, err := chainOb.Client.BlockNumber(context.TODO())
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s BlockNumber error", chainOb.chain)
-	//		return err
-	//	}
-	//	fromAddr := ethcommon.HexToAddress(config.TSS_TEST_ADDRESS)
-	//	toAddr := ethcommon.HexToAddress(config.POLYGON_MPI_ADDRESS)
-	//	res, err := chainOb.Client.CallContract(context.TODO(), ethereum.CallMsg{
-	//		From: fromAddr,
-	//		To:   &toAddr,
-	//		Data: input,
-	//	}, big.NewInt(0).SetUint64(bn))
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s CallContract error", chainOb.chain)
-	//		return err
-	//	}
-	//	output, err := chainOb.connectorAbi.Unpack("totalSupply", res)
-	//	if err != nil {
-	//		log.Err(err).Msgf("%s Unpack error", chainOb.chain)
-	//		return err
-	//	}
-	//	totalSupply := *connectorAbi.ConvertType(output[0], new(*big.Int)).(**big.Int)
-	//	//fmt.Printf("BSC: block %d: totalSupply %d\n", bn, totalSupply)
-	//	supply = totalSupply.String()
-	//} else {
-	//	log.Error().Msgf("chain not supported %s", chainOb.chain)
-	//	return fmt.Errorf("unsupported chain %s", chainOb.chain)
-	//}
-
-	_, err = ob.bridge.PostGasPrice(ob.chain, gasPrice.Uint64(), supply, blockNum)
-	if err != nil {
-		log.Err(err).Msg("PostGasPrice:")
-		return err
-	}
-
-	//bal, err := chainOb.Client.BalanceAt(context.TODO(), chainOb.Tss.Address(), nil)
-	//if err != nil {
-	//	log.Err(err).Msg("BalanceAt:")
-	//	return err
-	//}
-	//_, err = chainOb.bridge.PostGasBalance(chainOb.chain, bal.String(), blockNum)
-	//if err != nil {
-	//	log.Err(err).Msg("PostGasBalance:")
-	//	return err
-	//}
-	return nil
-}
-
-func (ob *ChainObserver) observeChain() error {
-	header, err := ob.Client.HeaderByNumber(context.Background(), nil)
-	if err != nil {
-		return err
-	}
-	counter, err := ob.GetPromCounter("rpc_getBlockByNumber_count")
-	if err != nil {
-		log.Error().Err(err).Msg("GetPromCounter:")
-	}
-	counter.Inc()
-
-	// "confirmed" current block number
-	confirmedBlockNum := header.Number.Uint64() - ob.confCount
-	// skip if no new block is produced.
-	if confirmedBlockNum <= ob.LastBlock {
-		return nil
-	}
-	toBlock := ob.LastBlock + config.MAX_BLOCKS_PER_PERIOD // read at most 10 blocks in one go
-	if toBlock >= confirmedBlockNum {
-		toBlock = confirmedBlockNum
-	}
-
-	topics[0] = []ethcommon.Hash{logZetaSentSignatureHash}
-
-	query := ethereum.FilterQuery{
-		Addresses: []ethcommon.Address{ethcommon.HexToAddress(ob.mpiAddress)},
-		FromBlock: big.NewInt(0).SetUint64(ob.LastBlock + 1), // LastBlock has been processed;
-		ToBlock:   big.NewInt(0).SetUint64(toBlock),
-		Topics:    topics,
-	}
-	//log.Debug().Msgf("signer %s block from %d to %d", chainOb.bridge.GetKeys().signerName, query.FromBlock, query.ToBlock)
-	ob.sampleLogger.Info().Msgf("%s current block %d, querying from %d to %d, %d blocks left to catch up, watching MPI address %s", ob.chain, header.Number.Uint64(), ob.LastBlock+1, toBlock, int(toBlock)-int(confirmedBlockNum), ethcommon.HexToAddress(ob.mpiAddress))
-
-	// Finally query the for the logs
-	logs, err := ob.Client.FilterLogs(context.Background(), query)
-	if err != nil {
-		return err
-	}
-	cnt, err := ob.GetPromCounter("rpc_getLogs_count")
-	if err != nil {
-		return err
-	}
-	cnt.Inc()
-
-	// Read in ABI
-	contractAbi := ob.connectorAbi
-
-	// Pull out arguments from logs
-	for _, vLog := range logs {
-		log.Info().Msgf("TxBlockNumber %d Transaction Hash: %s topic %s\n", vLog.BlockNumber, vLog.TxHash.Hex()[:6], vLog.Topics[0].Hex()[:6])
-		switch vLog.Topics[0].Hex() {
-		case logZetaSentSignatureHash.Hex():
-			vals, err := contractAbi.Unpack("ZetaSent", vLog.Data)
-			if err != nil {
-				log.Err(err).Msg("error unpacking ZetaMessageSendEvent")
-				continue
-			}
-			sender := vLog.Topics[1]
-			destChainID := vals[0].(*big.Int)
-			destContract := vals[1].([]byte)
-			zetaAmount := vals[2].(*big.Int)
-			gasLimit := vals[3].(*big.Int)
-			message := vals[4].([]byte)
-			zetaParams := vals[5].([]byte)
-
-			_ = zetaParams
-
-			metaHash, err := ob.bridge.PostSend(
-				ethcommon.HexToAddress(sender.Hex()).Hex(),
-				ob.chain.String(),
-				types.BytesToEthHex(destContract),
-				config.FindChainByID(destChainID),
-				zetaAmount.String(),
-				zetaAmount.String(),
-				base64.StdEncoding.EncodeToString(message),
-				vLog.TxHash.Hex(),
-				vLog.BlockNumber,
-				gasLimit.Uint64(),
-			)
-			if err != nil {
-				log.Err(err).Msg("error posting to meta core")
-				continue
-			}
-			log.Debug().Msgf("LockSend detected: PostSend metahash: %s", metaHash)
-		}
-	}
-
-	ob.LastBlock = toBlock
-	buf := make([]byte, binary.MaxVarintLen64)
-	n := binary.PutUvarint(buf, toBlock)
-	err = ob.db.Put([]byte(PosKey), buf[:n], nil)
-	if err != nil {
-		log.Error().Err(err).Msg("error writing toBlock to db")
-	}
-	return nil
-}
-
-// query ZetaCore about the last block that it has heard from a specific chain.
-// return 0 if not existent.
-func (ob *ChainObserver) getLastHeight() uint64 {
-	lastheight, err := ob.bridge.GetLastBlockHeightByChain(ob.chain)
-	if err != nil {
-		log.Warn().Err(err).Msgf("getLastHeight")
-		return 0
-	}
-	return lastheight.LastSendHeight
-}
-
-// query the base gas price for the block number bn.
-func (ob *ChainObserver) GetBaseGasPrice() *big.Int {
-	gasPrice, err := ob.Client.SuggestGasPrice(context.TODO())
-	if err != nil {
-		log.Err(err).Msg("GetBaseGasPrice")
-		return nil
-	}
-	return gasPrice
 }
 
 func (ob *ChainObserver) Stop() {
@@ -683,7 +337,7 @@ func (ob *ChainObserver) Stop() {
 // returns: isIncluded, isConfirmed, Error
 // If isConfirmed, it also post to ZetaCore
 func (ob *ChainObserver) IsSendOutTxProcessed(sendHash string, nonce int) (bool, bool, error) {
-	receipt, found := ob.nonceTx[nonce]
+	receipt, found := ob.outTXConfirmed[nonce]
 	if found && receipt.Status == 1 {
 		logs := receipt.Logs
 		for _, vLog := range logs {
@@ -701,7 +355,7 @@ func (ob *ChainObserver) IsSendOutTxProcessed(sendHash string, nonce int) (bool,
 					sendhash := vLog.Topics[3].Hex()
 					//var rxAddress string = ethcommon.HexToAddress(vLog.Topics[1].Hex()).Hex()
 					var mMint string = retval[1].(*big.Int).String()
-					metaHash, err := ob.bridge.PostReceiveConfirmation(
+					metaHash, err := ob.zetaClient.PostReceiveConfirmation(
 						sendhash,
 						vLog.TxHash.Hex(),
 						vLog.BlockNumber,
@@ -731,7 +385,7 @@ func (ob *ChainObserver) IsSendOutTxProcessed(sendHash string, nonce int) (bool,
 					log.Info().Msg("Confirmed! Sending PostConfirmation to zetacore...")
 					sendhash := vLog.Topics[3].Hex()
 					var mMint string = retval[2].(*big.Int).String()
-					metaHash, err := ob.bridge.PostReceiveConfirmation(
+					metaHash, err := ob.zetaClient.PostReceiveConfirmation(
 						sendhash,
 						vLog.TxHash.Hex(),
 						vLog.BlockNumber,
@@ -754,7 +408,7 @@ func (ob *ChainObserver) IsSendOutTxProcessed(sendHash string, nonce int) (bool,
 	} else if found && receipt.Status == 0 {
 		//FIXME: check nonce here by getTransaction RPC
 		log.Info().Msgf("Found (failed tx) sendHash %s on chain %s txhash %s", sendHash, ob.chain, receipt.TxHash.Hex())
-		zetaTxHash, err := ob.bridge.PostReceiveConfirmation(sendHash, receipt.TxHash.Hex(), receipt.BlockNumber.Uint64(), "", common.ReceiveStatus_Failed, ob.chain.String())
+		zetaTxHash, err := ob.zetaClient.PostReceiveConfirmation(sendHash, receipt.TxHash.Hex(), receipt.BlockNumber.Uint64(), "", common.ReceiveStatus_Failed, ob.chain.String())
 		if err != nil {
 			log.Error().Err(err).Msgf("PostReceiveConfirmation error in WatchTxHashWithTimeout; zeta tx hash %s", zetaTxHash)
 		}
@@ -765,8 +419,9 @@ func (ob *ChainObserver) IsSendOutTxProcessed(sendHash string, nonce int) (bool,
 	return false, false, fmt.Errorf("IsSendOutTxProcessed: error on chain %s", ob.chain)
 }
 
-// this function periodically checks all the txhash in potential outbound txs
 // FIXME: there's a chance that a txhash in OutTxChan may not deliver when Stop() is called
+// observeOutTx periodically checks all the txhash in potential outbound txs
+// If it is able to confirm one of txHashes in a outTXPending , it cleans it and saves information to local DB
 func (ob *ChainObserver) observeOutTx() {
 
 	ticker := time.NewTicker(12 * time.Second)
@@ -774,8 +429,8 @@ func (ob *ChainObserver) observeOutTx() {
 		select {
 		case <-ticker.C:
 			minNonce, maxNonce, err := ob.PurgeTxHashWatchList()
-			if len(ob.nonceTxHashesMap) > 0 {
-				log.Info().Msgf("chain %s outstanding nonce: %d; nonce range [%d,%d]", ob.chain, len(ob.nonceTxHashesMap), minNonce, maxNonce)
+			if len(ob.outTXPending) > 0 {
+				log.Info().Msgf("chain %s outstanding nonce: %d; nonce range [%d,%d]", ob.chain, len(ob.outTXPending), minNonce, maxNonce)
 			}
 			outTimeout := time.After(12 * time.Second)
 			if err == nil {
@@ -786,7 +441,7 @@ func (ob *ChainObserver) observeOutTx() {
 				//for nonce, txHashes := range ob.nonceTxHashesMap {
 				for nonce := minNonce; nonce <= maxNonce; nonce++ { // ensure lower nonce is queried first
 					ob.mu.Lock()
-					txHashes, found := ob.nonceTxHashesMap[nonce]
+					txHashes, found := ob.outTXPending[nonce]
 					txHashesCopy := txHashes
 					ob.mu.Unlock()
 					if !found {
@@ -804,11 +459,11 @@ func (ob *ChainObserver) observeOutTx() {
 							if err == nil && receipt != nil { // confirmed
 								log.Info().Msgf("observeOutTx: %s nonce %d, txHash %s confirmed", ob.chain, nonce, txHash)
 								ob.mu.Lock()
-								delete(ob.nonceTxHashesMap, nonce)
+								delete(ob.outTXPending, nonce)
 								if err = ob.db.Delete([]byte(NonceTxHashesKeyPrefix+fmt.Sprintf("%d", nonce)), nil); err != nil {
 									log.Error().Err(err).Msgf("PurgeTxHashWatchList: error deleting nonce %d tx hashes from db", nonce)
 								}
-								ob.nonceTx[nonce] = receipt
+								ob.outTXConfirmed[nonce] = receipt
 								value, err := receipt.MarshalJSON()
 								if err != nil {
 									log.Error().Err(err).Msgf("receipt marshal error %s", receipt.TxHash.Hex())
@@ -836,57 +491,6 @@ func (ob *ChainObserver) observeOutTx() {
 	}
 }
 
-// remove txhash from watch list which have no corresponding sendPending in zetacore.
-// returns the min/max nonce after purge
-func (ob *ChainObserver) PurgeTxHashWatchList() (int, int, error) {
-	purgedTxHashCount := 0
-	sends, err := ob.bridge.GetAllPendingSend()
-	if err != nil {
-		return 0, 0, err
-	}
-	pendingNonces := make(map[int]bool)
-	for _, send := range sends {
-		if send.Status == zetatypes.SendStatus_PendingRevert && send.SenderChain == ob.chain.String() {
-			pendingNonces[int(send.Nonce)] = true
-		} else if send.Status == zetatypes.SendStatus_PendingOutbound && send.ReceiverChain == ob.chain.String() {
-			pendingNonces[int(send.Nonce)] = true
-		}
-	}
-	tNow := time.Now()
-	ob.mu.Lock()
-	for nonce, _ := range ob.nonceTxHashesMap {
-		if _, found := pendingNonces[nonce]; !found {
-			txHashes := ob.nonceTxHashesMap[nonce]
-			delete(ob.nonceTxHashesMap, nonce)
-			if err = ob.db.Delete([]byte(NonceTxHashesKeyPrefix+fmt.Sprintf("%d", nonce)), nil); err != nil {
-				log.Error().Err(err).Msgf("PurgeTxHashWatchList: error deleting nonce %d tx hashes from db", nonce)
-			}
-			purgedTxHashCount++
-			log.Info().Msgf("PurgeTxHashWatchList: chain %s nonce %d removed", ob.chain, nonce)
-			ob.fileLogger.Info().Msgf("PurgeTxHashWatchList: chain %s nonce %d removed txhashes %v", ob.chain, nonce, txHashes)
-		}
-	}
-	ob.mu.Unlock()
-	if purgedTxHashCount > 0 {
-		log.Info().Msgf("PurgeTxHashWatchList: chain %s purged %d txhashes in %v", ob.chain, purgedTxHashCount, time.Since(tNow))
-	}
-	minNonce, maxNonce := -1, 0
-	if len(pendingNonces) > 0 {
-		for nonce, _ := range pendingNonces {
-			if minNonce == -1 {
-				minNonce = nonce
-			}
-			if nonce < minNonce {
-				minNonce = nonce
-			}
-			if nonce > maxNonce {
-				maxNonce = nonce
-			}
-		}
-	}
-	return minNonce, maxNonce, nil
-}
-
 // return the status of txHash
 // receipt nil, err non-nil: txHash not found
 // receipt non-nil, err non-nil: txHash found but not confirmed
@@ -896,7 +500,7 @@ func (ob *ChainObserver) queryTxByHash(txHash string, nonce int) (*ethtypes.Rece
 	//defer func() { log.Info().Msgf("queryTxByHash elapsed: %s", time.Since(timeStart)) }()
 	ctxt, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	defer cancel()
-	receipt, err := ob.Client.TransactionReceipt(ctxt, ethcommon.HexToHash(txHash))
+	receipt, err := ob.EvmClient.TransactionReceipt(ctxt, ethcommon.HexToHash(txHash))
 	if err != nil {
 		if err != ethereum.NotFound {
 			log.Warn().Err(err).Msgf("%s %s TransactionReceipt err", ob.chain, txHash)
@@ -912,27 +516,5 @@ func (ob *ChainObserver) queryTxByHash(txHash string, nonce int) (*ethtypes.Rece
 			log.Info().Msgf("%s TransactionReceipt %s nonce %d mined and confirmed, and it's successful", ob.chain, txHash, nonce)
 		}
 		return receipt, nil
-	}
-}
-
-func (ob *ChainObserver) AddTxHashToWatchList(txHash string, nonce int, sendHash string) {
-	outTx := OutTx{
-		TxHash:   txHash,
-		Nonce:    nonce,
-		SendHash: sendHash,
-	}
-
-	if outTx.TxHash != "" { // TODO: this seems unnecessary
-		ob.mu.Lock()
-		ob.nonceTxHashesMap[outTx.Nonce] = append(ob.nonceTxHashesMap[outTx.Nonce], outTx.TxHash)
-		ob.mu.Unlock()
-		key := []byte(NonceTxHashesKeyPrefix + fmt.Sprintf("%d", outTx.Nonce))
-		value := []byte(strings.Join(ob.nonceTxHashesMap[outTx.Nonce], ","))
-		if err := ob.db.Put(key, value, nil); err != nil {
-			log.Error().Err(err).Msgf("AddTxHashToWatchList: error adding nonce %d tx hashes to db", outTx.Nonce)
-		}
-
-		log.Info().Msgf("add %s nonce %d TxHash watch list length: %d", ob.chain, outTx.Nonce, len(ob.nonceTxHashesMap[outTx.Nonce]))
-		ob.fileLogger.Info().Msgf("add %s nonce %d TxHash watch list length: %d", ob.chain, outTx.Nonce, len(ob.nonceTxHashesMap[outTx.Nonce]))
 	}
 }
