+++ zetaclient/evm_chain_utils.go
@@ -1,0 +1,137 @@
+package zetaclient
+
+import (
+	"context"
+	"encoding/base64"
+	"encoding/binary"
+	"github.com/ethereum/go-ethereum"
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/rs/zerolog/log"
+	"github.com/zeta-chain/zetacore/zetaclient/config"
+	"github.com/zeta-chain/zetacore/zetaclient/types"
+	"math/big"
+)
+
+func (ob *ChainObserver) ExternalChainWatcher() {
+	// At each tick, query the mpiAddress
+	for {
+		select {
+		case <-ob.ticker.C:
+			err := ob.observeInTX()
+			if err != nil {
+				log.Err(err).Msg("observeInTX error on " + ob.chain.String())
+				continue
+			}
+		case <-ob.stop:
+			log.Info().Msg("ExternalChainWatcher stopped")
+			return
+		}
+	}
+}
+
+func (ob *ChainObserver) observeInTX() error {
+	header, err := ob.EvmClient.HeaderByNumber(context.Background(), nil)
+	if err != nil {
+		return err
+	}
+	counter, err := ob.GetPromCounter("rpc_getBlockByNumber_count")
+	if err != nil {
+		log.Error().Err(err).Msg("GetPromCounter:")
+	}
+	counter.Inc()
+
+	// "confirmed" current block number
+	confirmedBlockNum := header.Number.Uint64() - ob.confCount
+	// skip if no new block is produced.
+	if confirmedBlockNum <= ob.LastBlock {
+		return nil
+	}
+	toBlock := ob.LastBlock + config.MAX_BLOCKS_PER_PERIOD // read at most 10 blocks in one go
+	if toBlock >= confirmedBlockNum {
+		toBlock = confirmedBlockNum
+	}
+
+	topics[0] = []ethcommon.Hash{logZetaSentSignatureHash}
+
+	query := ethereum.FilterQuery{
+		Addresses: []ethcommon.Address{ethcommon.HexToAddress(ob.mpiAddress)},
+		FromBlock: big.NewInt(0).SetUint64(ob.LastBlock + 1), // LastBlock has been processed;
+		ToBlock:   big.NewInt(0).SetUint64(toBlock),
+		Topics:    topics,
+	}
+	//log.Debug().Msgf("signer %s block from %d to %d", chainOb.zetaClient.GetKeys().signerName, query.FromBlock, query.ToBlock)
+	ob.sampleLogger.Info().Msgf("%s current block %d, querying from %d to %d, %d blocks left to catch up, watching MPI address %s", ob.chain, header.Number.Uint64(), ob.LastBlock+1, toBlock, int(toBlock)-int(confirmedBlockNum), ethcommon.HexToAddress(ob.mpiAddress))
+
+	// Finally query the for the logs
+	logs, err := ob.EvmClient.FilterLogs(context.Background(), query)
+	if err != nil {
+		return err
+	}
+	cnt, err := ob.GetPromCounter("rpc_getLogs_count")
+	if err != nil {
+		return err
+	}
+	cnt.Inc()
+
+	// Read in ABI
+	contractAbi := ob.connectorAbi
+
+	// Pull out arguments from logs
+	for _, vLog := range logs {
+		log.Info().Msgf("TxBlockNumber %d Transaction Hash: %s topic %s\n", vLog.BlockNumber, vLog.TxHash.Hex()[:6], vLog.Topics[0].Hex()[:6])
+		switch vLog.Topics[0].Hex() {
+		case logZetaSentSignatureHash.Hex():
+			vals, err := contractAbi.Unpack("ZetaSent", vLog.Data)
+			if err != nil {
+				log.Err(err).Msg("error unpacking ZetaMessageSendEvent")
+				continue
+			}
+			sender := vLog.Topics[1]
+			destChainID := vals[0].(*big.Int)
+			destContract := vals[1].([]byte)
+			zetaAmount := vals[2].(*big.Int)
+			gasLimit := vals[3].(*big.Int)
+			message := vals[4].([]byte)
+			zetaParams := vals[5].([]byte)
+
+			_ = zetaParams
+
+			metaHash, err := ob.zetaClient.PostSend(
+				ethcommon.HexToAddress(sender.Hex()).Hex(),
+				ob.chain.String(),
+				types.BytesToEthHex(destContract),
+				config.FindChainByID(destChainID),
+				zetaAmount.String(),
+				zetaAmount.String(),
+				base64.StdEncoding.EncodeToString(message),
+				vLog.TxHash.Hex(),
+				vLog.BlockNumber,
+				gasLimit.Uint64(),
+			)
+			if err != nil {
+				log.Err(err).Msg("error posting to meta core")
+				continue
+			}
+			log.Debug().Msgf("LockSend detected: PostSend metahash: %s", metaHash)
+		}
+	}
+
+	ob.LastBlock = toBlock
+	buf := make([]byte, binary.MaxVarintLen64)
+	n := binary.PutUvarint(buf, toBlock)
+	err = ob.db.Put([]byte(PosKey), buf[:n], nil)
+	if err != nil {
+		log.Error().Err(err).Msg("error writing toBlock to db")
+	}
+	return nil
+}
+
+// query the base gas price for the block number bn.
+func (ob *ChainObserver) GetBaseGasPrice() *big.Int {
+	gasPrice, err := ob.EvmClient.SuggestGasPrice(context.TODO())
+	if err != nil {
+		log.Err(err).Msg("GetBaseGasPrice")
+		return nil
+	}
+	return gasPrice
+}
