+++ zetaclient/voter_test.go
@@ -34,7 +34,7 @@ func (s *VoterSuite) SetUpTest(c *C) {
 	chainHomeFoler := filepath.Join(homeDir, ".zetacore")
 	c.Logf("chain home dir: %s", chainHomeFoler)
 
-	// first signer & bridge
+	// first signer & zetaClient
 	// alice is the default user created by Starport chain serve
 	{
 		signerName := "alice"
@@ -57,7 +57,7 @@ func (s *VoterSuite) SetUpTest(c *C) {
 		s.bridge1 = bridge
 	}
 
-	// second signer & bridge
+	// second signer & zetaClient
 	// alice is the default user created by Starport chain serve
 	{
 		signerName := "bob"
